requirements.txt
gunicorn==23.0.0
accelerate==1.5.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
construct==2.5.3
defusedxml==0.7.1
distro==1.9.0
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.0
h11==0.14.0
h5py==3.13.0
haystack-ai==2.11.2
haystack-experimental==0.8.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langdetect==1.0.9
lazy_imports==0.4.0
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
openai==1.68.2
opencv-python==4.11.0.86
packaging==24.2
pefile==2024.8.26
pillow==11.1.0
posthog==3.23.0
protobuf==6.30.1
psutil==7.0.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-doctr==0.11.0
python-ptrace==0.9.9
PyYAML==6.0.2
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scipy==1.15.2
setuptools==79.0.1 
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.1
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
